Time: 7 ms (91.34%) | Memory: 35 MB (53.93%) - LeetSync

+++ 1056-capacity-to-ship-packages-within-d-days/capacity-to-ship-packages-within-d-days.cpp
@@ -9,7 +9,7 @@ public:
             low = max(low,w[i]);
             high+=w[i];
         }
-        long ans =high;
+        int ans =high;
             while(low<=high){
              mid = low+(high-low)/2;
